Renamed typeconverter package to type.converter
