+++ difuse-core/src/main/java/com/kaba4cow/difuse/core/type/converter/impl/primitive/CharTypeConverter.java
@@ -1,8 +1,8 @@
-package com.kaba4cow.difuse.core.typeconverter.impl.primitive;
+package com.kaba4cow.difuse.core.type.converter.impl.primitive;
 
-import com.kaba4cow.difuse.core.typeconverter.PrimitiveTypeConverter;
-import com.kaba4cow.difuse.core.typeconverter.TypeConverter;
-import com.kaba4cow.difuse.core.typeconverter.TypeConverterException;
+import com.kaba4cow.difuse.core.type.converter.PrimitiveTypeConverter;
+import com.kaba4cow.difuse.core.type.converter.TypeConverter;
+import com.kaba4cow.difuse.core.type.converter.TypeConverterException;
 
 public class CharTypeConverter implements TypeConverter<Character>, PrimitiveTypeConverter {
