+++ difuse-core/src/main/java/com/kaba4cow/difuse/core/type/converter/support/TypeConverterInitializer.java
@@ -1,4 +1,4 @@
-package com.kaba4cow.difuse.core.typeconverter.support;
+package com.kaba4cow.difuse.core.type.converter.support;
 
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -10,7 +10,7 @@ import com.kaba4cow.difuse.core.annotation.dependency.Provided;
 import com.kaba4cow.difuse.core.annotation.system.SystemBean;
 import com.kaba4cow.difuse.core.context.ContextScanner;
 import com.kaba4cow.difuse.core.context.source.support.ContextSourceRegistry;
-import com.kaba4cow.difuse.core.typeconverter.TypeConverter;
+import com.kaba4cow.difuse.core.type.converter.TypeConverter;
 import com.kaba4cow.difuse.core.util.LoggingTimer;
 
 @SystemBean
