+++ difuse-core/src/main/java/com/kaba4cow/difuse/core/type/converter/support/TypeConverterRegistrar.java
@@ -1,11 +1,11 @@
-package com.kaba4cow.difuse.core.typeconverter.support;
+package com.kaba4cow.difuse.core.type.converter.support;
 
 import java.lang.reflect.Constructor;
 
 import com.kaba4cow.difuse.core.DifuseException;
 import com.kaba4cow.difuse.core.annotation.dependency.Provided;
 import com.kaba4cow.difuse.core.annotation.system.SystemBean;
-import com.kaba4cow.difuse.core.typeconverter.TypeConverter;
+import com.kaba4cow.difuse.core.type.converter.TypeConverter;
 import com.kaba4cow.difuse.core.util.reflections.ConstructorScanner;
 
 @SystemBean
