+++ difuse-core/src/main/java/com/kaba4cow/difuse/core/type/converter/support/converter/ArrayConverter.java
@@ -1,11 +1,11 @@
-package com.kaba4cow.difuse.core.typeconverter.support.converter;
+package com.kaba4cow.difuse.core.type.converter.support.converter;
 
 import java.lang.reflect.Array;
 import java.lang.reflect.Type;
 import java.util.Collection;
 
-import com.kaba4cow.difuse.core.typeconverter.TypeConverterException;
-import com.kaba4cow.difuse.core.typeconverter.support.GlobalTypeConverter;
+import com.kaba4cow.difuse.core.type.converter.TypeConverterException;
+import com.kaba4cow.difuse.core.type.converter.support.GlobalTypeConverter;
 
 public class ArrayConverter {
