+++ difuse-core/src/main/java/com/kaba4cow/difuse/core/type/converter/support/converter/MapConverter.java
@@ -1,12 +1,12 @@
-package com.kaba4cow.difuse.core.typeconverter.support.converter;
+package com.kaba4cow.difuse.core.type.converter.support.converter;
 
 import java.lang.reflect.Type;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 
-import com.kaba4cow.difuse.core.typeconverter.TypeConverterException;
-import com.kaba4cow.difuse.core.typeconverter.support.GlobalTypeConverter;
+import com.kaba4cow.difuse.core.type.converter.TypeConverterException;
+import com.kaba4cow.difuse.core.type.converter.support.GlobalTypeConverter;
 
 public class MapConverter {
