+++ difuse-core/src/main/java/com/kaba4cow/difuse/core/type/converter/support/converter/OptionalConverter.java
@@ -1,10 +1,10 @@
-package com.kaba4cow.difuse.core.typeconverter.support.converter;
+package com.kaba4cow.difuse.core.type.converter.support.converter;
 
 import java.lang.reflect.Type;
 import java.util.Objects;
 import java.util.Optional;
 
-import com.kaba4cow.difuse.core.typeconverter.support.GlobalTypeConverter;
+import com.kaba4cow.difuse.core.type.converter.support.GlobalTypeConverter;
 
 public class OptionalConverter {
